--- hook-bootkit/go.mod
@@ -1,33 +1,0 @@
-module github.com/tinkerbell/hook/hook-bootkit
-
-go 1.17
-
-require (
-	github.com/cenkalti/backoff/v4 v4.2.1
-	github.com/docker/docker v24.0.4+incompatible
-	github.com/go-logr/logr v1.2.4
-	github.com/go-logr/zerologr v1.2.3
-	github.com/rs/zerolog v1.30.0
-)
-
-require (
-	github.com/Microsoft/go-winio v0.6.1 // indirect
-	github.com/docker/distribution v2.8.2+incompatible // indirect
-	github.com/docker/go-connections v0.4.0 // indirect
-	github.com/docker/go-units v0.5.0 // indirect
-	github.com/gogo/protobuf v1.3.2 // indirect
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/moby/term v0.5.0 // indirect
-	github.com/morikuni/aec v1.0.0 // indirect
-	github.com/opencontainers/go-digest v1.0.0 // indirect
-	github.com/opencontainers/image-spec v1.0.2 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/stretchr/testify v1.8.4 // indirect
-	golang.org/x/mod v0.8.0 // indirect
-	golang.org/x/net v0.12.0 // indirect
-	golang.org/x/sys v0.10.0 // indirect
-	golang.org/x/time v0.3.0 // indirect
-	golang.org/x/tools v0.6.0 // indirect
-	gotest.tools/v3 v3.5.0 // indirect
-)
