--- hook-docker/go.mod
@@ -1,3 +1,0 @@
-module github.com/tinkerbell/hook/hook-docker
-
-go 1.18
