--- hook-mdev/Dockerfile
@@ -1,8 +1,0 @@
-FROM alpine
-
-USER root:root
-
-RUN apk add --no-cache mdev-conf
-
-CMD ["mdev", "-v", "-df"]
-
