+++ images/hook-docker/go.mod
@@ -1,0 +1,3 @@
+module github.com/tinkerbell/hook/hook-docker
+
+go 1.18
