+++ images/hook-mdev/Dockerfile
@@ -1,0 +1,8 @@
+FROM alpine
+
+USER root:root
+
+RUN apk add --no-cache mdev-conf
+
+CMD ["mdev", "-v", "-df"]
+
