+++ kernel/Dockerfile
@@ -64,6 +64,8 @@ ARG KERNEL_OUTPUT_IMAGE
 
 RUN mkdir /out
 
+RUN sed -i 's/#define COMMAND_LINE_SIZE 2048/#define COMMAND_LINE_SIZE 4096/' arch/x86/include/asm/setup.h
+
 # Kernel build. ENVs in previous stages are inherited; thus ARCH, CROSS_COMPILE, KCFLAGS, KBUILD_BUILD_USER, KBUILD_BUILD_HOST are available
 RUN set -x && \
     echo "Cross compiler: ${CROSS_COMPILE}" && \
